% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_all}
\alias{check_all}
\title{Check that all rows match a condition}
\usage{
check_all(condition, check_name = NULL)
}
\arguments{
\item{condition}{a condition}

\item{check_name}{the name of the check}
}
\value{
a data check
}
\description{
Check that all rows match a condition
}
\examples{

pass <- check_all(mpg < 34)
pass(mtcars)
}
