% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_no_duplicates}
\alias{check_no_duplicates}
\title{Check that a column is not duplicated}
\usage{
check_no_duplicates(col)
}
\arguments{
\item{col}{column to check}
}
\value{
a dqc object
}
\description{
Check that a column is not duplicated
}
