% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_no_na}
\alias{check_no_na}
\title{Check that a column has no NA values}
\usage{
check_no_na(col)
}
\arguments{
\item{col}{column}
}
\value{
a data check object
}
\description{
Check that a column has no NA values
}
