% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_values_in}
\alias{check_values_in}
\title{Check that a column's values are contained in some set}
\usage{
check_values_in(col, values)
}
\arguments{
\item{col}{an unquoted column}

\item{values}{values to check against}
}
\value{
a data check object
}
\description{
Check that a column's values are contained in some set
}
