% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chex-package.R
\docType{package}
\name{chex-package}
\alias{chex}
\alias{chex-package}
\title{chex: Simple Data Integrity Checks}
\description{
Check that a dataset passes data integrity checks.
}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
