% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chex-package.R
\docType{package}
\name{dqc-package}
\alias{dqc}
\alias{dqc-package}
\title{dqc: Simple Data Quality Checks}
\description{
Check that a dataset passes data integrity checks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/chex}
  \item Report bugs at \url{https://github.com/colin-fraser/chex/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
