% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqc.R
\name{dqc}
\alias{dqc}
\alias{format.dqc}
\alias{print.dqc}
\title{Create a Data Quality Check}
\usage{
dqc(check_name, f, expected_output = TRUE)

\method{format}{dqc}(x, ...)

\method{print}{dqc}(x, ...)
}
\arguments{
\item{check_name}{the name of the check}

\item{f}{the check function}

\item{expected_output}{expected output of the check}

\item{x}{a dqc object}

\item{...}{unused}
}
\value{
a dqc object
}
\description{
Create a Data Quality Check
}
