% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqc_result.R
\name{dqc_result}
\alias{dqc_result}
\alias{format.dqc_result}
\alias{print.dqc_result}
\title{The result of a Data Quality Check}
\usage{
dqc_result(check_name, expected_output, check_passed, ...)

\method{format}{dqc_result}(x, ...)

\method{print}{dqc_result}(x, ...)
}
\arguments{
\item{check_name}{the name of the check}

\item{expected_output}{the expected output of the check}

\item{check_passed}{did the check pass?}

\item{...}{unused}

\item{x}{a dqc_result object}
}
\value{
a dqc_result object
}
\description{
The result of a Data Quality Check
}
