% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqc_results.R
\name{dqc_results}
\alias{dqc_results}
\title{Data Quality Check Suite Results}
\usage{
dqc_results(x)
}
\arguments{
\item{x}{a list of data quality check results}
}
\value{
A dqc_results object
}
\description{
Data Quality Check Suite Results
}
