% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqc_suite.R
\name{dqc_suite}
\alias{dqc_suite}
\title{Create a Data Quality Check Suite}
\usage{
dqc_suite(...)
}
\arguments{
\item{...}{a list of data quality checks}
}
\value{
a Data Quality Check Suite object
}
\description{
Create a Data Quality Check Suite
}
