% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_check.R
\name{quick_check}
\alias{quick_check}
\title{Quick Checks}
\usage{
quick_check(
  .data,
  ...,
  .behavior = c("inform", "warn", "stop"),
  .quiet_pass = FALSE
)
}
\arguments{
\item{.data}{a data.frame or similar}

\item{...}{data quality checks}

\item{.behavior}{what to do if checks fail}

\item{.quiet_pass}{do you want a message if all checks pass?}
}
\value{
.data, unchanged
}
\description{
Quick Checks
}
